+++ CHANGES.rst
@@ -50,6 +50,9 @@ New Features
 
 - ``astropy.io.fits``
 
+  - Included a new command-line script called ``fitsheader`` to display the
+    header(s) of a FITS file from the command line. [#2092]
+
 - ``astropy.io.misc``
 
 - ``astropy.io.registry``